requirements.txt
# requirements.txt

# Drake
drake
manipulation

# Visualization and Math
numpy
meshcat
matplotlib
scipy

# 3D Geometry
trimesh

# Testing
pytest
pytest-cov

# Debugging
ipdb
typer

# Graphs
networkx